refactor: rename ArticleViewHisAsyncHandler
